Try to only emit message if we open a new file
Only emit message if we open an output file with 0 file size, which is usually
a new file.

This avoids noise in out-of-diskspace situations when the error recovery
kicks in and does close/open on the output file (it's done this way to
cope with panicky sysadmins moving away old log files to free up space).